pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "aurora-app-player"
version = "0.0.1"
authors = [
    { name="Evgeniy Mazalov", email="guyboe@gmail.com" },
]
description = "Add play audio via configured player compatibility to aurora voice assistant"
readme = "README.md"
requires-python = ">=3.10"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

dependencies = [
    "typer==0.9.0",
    "kombu==5.2.4",
    "pydantic-another-config",
]

[project.urls]
"Homepage" = "https://github.com/guybo/aurora-app-player"
"Bug Tracker" = "https://github.com/guybo/aurora-app-player/issues"
